- Starter Negamax algorithm

+++ Juego IA/Assets/Scripts/ScoringMove.cs
@@ -1,9 +1,9 @@
 ﻿public class ScoringMove
 {
     public Tile tile;
-    public short score;
+    public int score;
 
-    public ScoringMove(short _score, Tile _tile)
+    public ScoringMove(int _score, Tile _tile)
     {
         tile = _tile;
         score = _score;
